09-seguridad/01-jwt/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
bcrypt
